Add files via upload
Amazed?, a maze solver bot developed using Arduino that aims to traverse the optimal path from the source to the destination point. Going crazy about solving the mazes? Give it to the BOT n see how it works it out…Uses the left-hand rule algorithm and efficiency has been captured in the algorithm by reducing the redundant and erroneous conditions caused due to the possible permutations and combinations of the line array sensors.